Moving parallel sessions around on Thursday and Friday
Geometric Group Theory is now using Port McNeill on Thursday so
Mathematics of Information shifts to Port Alberni and New trends in
Geometry and Mathematical Physics moves to Orca. On Friday there is no
Geometric Group Theory session, but we keep the same room assignments as
Thursday for Mathematics of Information (Port Alberni) and New Trends in
Geometry and Mathematical Physics (Orca).

Signed-off-by: Ian Allison <iana@pims.math.ca>